Move user CUD operations (not R) from API server to auth server
